Client Typescript Migration
- 필요 모듈 설치, 클라이언트 ts 세팅
- Client 모든 모듈 typescript migration
- Server Vibe 데이터 타입 Shared 폴더로 이동
