+++ packages/client/package.json
@@ -19,6 +19,8 @@
     "nuxt": "^2.0.0"
   },
   "devDependencies": {
+    "@nuxt/types": "^2.15.8",
+    "@nuxt/typescript-build": "^2.1.0",
     "@nuxtjs/vuetify": "^1.11.3"
   },
   "resolutions": {
