+++ packages/server/utils/resolveModulePath.ts
@@ -2,3 +2,4 @@ import * as ModuleAlias from 'module-alias';
 import * as path from 'path';
 
 ModuleAlias.addAliases({ '#': path.resolve(__dirname, '..', '..', 'shared') });
+ModuleAlias.addAliases({ '@': path.resolve(__dirname, '..') });
